# doc/api.yaml
openapi: 3.0.0
info:
  description: WASAText app backend REST API.
  title: WASAText API
  version: 1.0.0
  contact:
    name: Francesco Calzona
    email: calzona.2046920@studenti.uniroma1.it
    url: www.placeholder.com
  license:
    name: ECL-2.0
    url: https://opensource.org/licenses/ECL-2.0
servers:
    - url: http://wasatext.com/v1
...
paths:
...
  /settings/username:
    put:
      tags: ["user settings"]
      summary: Sets username
      description: |-
        It sets the username of the logged in user.
        This username will be showed to other people through messages sent,
        and it will be used to find them. 
      operationId: setMyUserName
      requestBody:
        description: username
        content:
          application/json:
            schema:
              type: object
              description: username
              properties:
                value:
                  $ref: '#/components/schemas/Username'
            example:
              value: Maria
        required: true
      responses:
        '400':
          description: Username parsed incorrectly or not valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 400
                message: The username is not valid
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '202':
          description: Username updated successfully
        '403':
          description: Username already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 403
                message: The username tried out is already in use
        '500':
            description: Internal server error
            content:
              application/json:
                schema:
                    $ref : '#/components/schemas/Error'
                example:
                  code: 500
                  message: Internal server error
...
components:
  schemas:
    Username:
      title: username
      description: This object represents a username
      type: string
      example: Maria
      pattern: ^\w{3,16}$
      minLength: 3
      maxLength: 16
...
    Error:
      type: object
      description: A generic error message
      properties:
        code:
          description: HTTP error code
          type: number
          example: 401
        message:
          description: error message
          type: string
          pattern: ^.{1,300}$
          minLength: 1
          maxLength: 300
          example: UnauthorizedError
      required:
        - code
        - message
      example:
        code: 401
        message: UnauthorizedError
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 401
            message: UnauthorizedError 
...
  securitySchemes:
    userID: # arbitrary name for the security scheme
      type: http
      description: The logging token of a single user
      scheme: bearer
      bearerFormat: JWT  # optional, for documentation purposes only
security:
  - userID: []